Update log4j package names
