+++ instrumentation/log4j/log4j-context-data/log4j-context-data-2.17/library-autoconfigure/src/main/resources/META-INF/services/org.apache.logging.log4j.core.util.ContextDataProvider
@@ -1,1 +1,1 @@
-io.opentelemetry.instrumentation.log4j.v2_16.OpenTelemetryContextDataProvider
+io.opentelemetry.instrumentation.log4j.contextdata.v2_17.OpenTelemetryContextDataProvider
