+++ instrumentation/log4j/log4j-context-data/log4j-context-data-2.7/javaagent/build.gradle.kts
@@ -16,7 +16,7 @@ dependencies {
 
   testInstrumentation(project(":instrumentation:log4j:log4j-context-data:log4j-context-data-2.17:javaagent"))
 
-  testImplementation(project(":instrumentation:log4j:log4j-context-data:log4j-context-data-2-common:testing"))
+  testImplementation(project(":instrumentation:log4j:log4j-context-data:log4j-context-data-common:testing"))
 
   latestDepTestLibrary("org.apache.logging.log4j:log4j-core:2.16.+") // see log4j-context-data-2.17
 }
