--- instrumentation/log4j/log4j-context-data/log4j-context-data-2.7/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/log4j/v2_7/ContextDataInjectorFactoryInstrumentation.java
@@ -1,49 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.log4j.v2_7;
-
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isPublic;
-import static net.bytebuddy.matcher.ElementMatchers.isStatic;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.returns;
-
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.implementation.bytecode.assign.Assigner;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.apache.logging.log4j.core.ContextDataInjector;
-
-public class ContextDataInjectorFactoryInstrumentation implements TypeInstrumentation {
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("org.apache.logging.log4j.core.impl.ContextDataInjectorFactory");
-  }
-
-  @Override
-  public void transform(TypeTransformer transformer) {
-    transformer.applyAdviceToMethod(
-        isMethod()
-            .and(isPublic())
-            .and(isStatic())
-            .and(named("createInjector"))
-            .and(returns(named("org.apache.logging.log4j.core.ContextDataInjector"))),
-        this.getClass().getName() + "$CreateInjectorAdvice");
-  }
-
-  @SuppressWarnings("unused")
-  public static class CreateInjectorAdvice {
-
-    @Advice.OnMethodExit(suppress = Throwable.class)
-    public static void onExit(
-        @Advice.Return(typing = Assigner.Typing.DYNAMIC, readOnly = false)
-            ContextDataInjector injector) {
-      injector = new SpanDecoratingContextDataInjector(injector);
-    }
-  }
-}
