+++ instrumentation/log4j/log4j-mdc-1.2/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/log4j/mdc/v1_2/LoggingEventInstrumentation.java
@@ -1,0 +1,85 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.log4j.mdc.v1_2;
+
+import static io.opentelemetry.instrumentation.api.log.LoggingContextConstants.SPAN_ID;
+import static io.opentelemetry.instrumentation.api.log.LoggingContextConstants.TRACE_FLAGS;
+import static io.opentelemetry.instrumentation.api.log.LoggingContextConstants.TRACE_ID;
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
+import static net.bytebuddy.matcher.ElementMatchers.isPublic;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
+import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
+
+import io.opentelemetry.api.trace.SpanContext;
+import io.opentelemetry.context.Context;
+import io.opentelemetry.instrumentation.api.field.VirtualField;
+import io.opentelemetry.javaagent.bootstrap.Java8BytecodeBridge;
+import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
+import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
+import net.bytebuddy.asm.Advice;
+import net.bytebuddy.description.type.TypeDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+import org.apache.log4j.spi.LoggingEvent;
+
+public class LoggingEventInstrumentation implements TypeInstrumentation {
+  @Override
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return named("org.apache.log4j.spi.LoggingEvent");
+  }
+
+  @Override
+  public void transform(TypeTransformer transformer) {
+    transformer.applyAdviceToMethod(
+        isMethod()
+            .and(isPublic())
+            .and(named("getMDC"))
+            .and(takesArguments(1))
+            .and(takesArgument(0, String.class)),
+        LoggingEventInstrumentation.class.getName() + "$GetMdcAdvice");
+  }
+
+  @SuppressWarnings("unused")
+  public static class GetMdcAdvice {
+
+    @Advice.OnMethodExit(suppress = Throwable.class)
+    public static void onExit(
+        @Advice.This LoggingEvent event,
+        @Advice.Argument(0) String key,
+        @Advice.Return(readOnly = false) Object value) {
+      if (TRACE_ID.equals(key) || SPAN_ID.equals(key) || TRACE_FLAGS.equals(key)) {
+        if (value != null) {
+          // Assume already instrumented event if traceId/spanId/sampled is present.
+          return;
+        }
+
+        Context context = VirtualField.find(LoggingEvent.class, Context.class).get(event);
+        if (context == null) {
+          return;
+        }
+
+        SpanContext spanContext = Java8BytecodeBridge.spanFromContext(context).getSpanContext();
+        if (!spanContext.isValid()) {
+          return;
+        }
+
+        switch (key) {
+          case TRACE_ID:
+            value = spanContext.getTraceId();
+            break;
+          case SPAN_ID:
+            value = spanContext.getSpanId();
+            break;
+          case TRACE_FLAGS:
+            value = spanContext.getTraceFlags().asHex();
+            break;
+          default:
+            // do nothing
+        }
+      }
+    }
+  }
+}
