+++ settings.gradle.kts
@@ -300,7 +300,7 @@ include(":instrumentation:log4j:log4j-mdc-1.2:javaagent")
 include(":instrumentation:log4j:log4j-context-data:log4j-context-data-2.7:javaagent")
 include(":instrumentation:log4j:log4j-context-data:log4j-context-data-2.17:javaagent")
 include(":instrumentation:log4j:log4j-context-data:log4j-context-data-2.17:library-autoconfigure")
-include(":instrumentation:log4j:log4j-context-data:log4j-context-data-2-common:testing")
+include(":instrumentation:log4j:log4j-context-data:log4j-context-data-common:testing")
 include(":instrumentation:log4j:log4j-appender-2.17:javaagent")
 include(":instrumentation:log4j:log4j-appender-2.17:library")
 include(":instrumentation:logback:logback-appender-1.0:javaagent")
